Refactor: 불필요한 코드 정리, clock있는 생성자 추가, headername활용

+++ elastic-search/src/main/java/store/_0982/elasticsearch/application/dto/GroupPurchaseDocumentInfo.java
@@ -11,7 +11,7 @@ public record GroupPurchaseDocumentInfo(
         Integer maxQuantity,
         String title,
         String description,
-        Long discountedPrice,
+        Integer discountedPrice,
         String status,
         String startDate,
         String endDate,
