+++ elastic-search/src/main/java/store/_0982/elasticsearch/presentation/dto/GroupPurchaseDocumentRequest.java
@@ -13,7 +13,7 @@ public record GroupPurchaseDocumentRequest(
         Integer maxQuantity,
         String title,
         String description,
-        Long discountedPrice,
+        Integer discountedPrice,
         String status,
         String startDate,
         String endDate,
